server/requirements.txt
# FastAPI and web server
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
python-multipart==0.0.6
email-validator==2.1.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Database - Tortoise ORM
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.28.0